mid refactor, refactoring schemas for user routes [skip ci]
