+++ apps/leaderboard-backend/src/validation/guildSchema.ts
@@ -1,7 +1,7 @@
 import type { Address } from "@happy.tech/common"
 import { z } from "@hono/zod-openapi"
 import { isHex } from "viem"
-import type { GuildTableId, UserTableId } from "../../db/types"
+import type { GuildTableId, UserTableId } from "../db/types"
 
 // Guild API response schema (for GET, POST responses)
 export const GuildResponseSchema = z
