+++ apps/leaderboard-backend/src/validation/users/index.ts
@@ -1,0 +1,3 @@
+export * from "./userRouteDescriptions"
+export * from "./userRouteValidations"
+export * from "./userSchemas"
